WIP: improve code like review comments and add plugin for getList  method
