+++ app/code/Magento/InventorySales/etc/extension_attributes.xml
@@ -6,7 +6,7 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Api/etc/extension_attributes.xsd">
-    <extension_attributes for="Magento\IventoryApi\Api\Data\StockInterface">
+    <extension_attributes for="Magento\InventoryApi\Api\Data\StockInterface">
         <attribute code="sales_channels" type="Magento\IventorySalesApi\Api\Data\SalesChannelInterface[]" />
     </extension_attributes>
 </config>
